feat: 添加消息模块，使用RabbitMQ实现消息异步解耦

+++ backend/user-service/user-server/src/main/java/ink/whi/user/model/dto/UserStatisticInfoDTO.java
@@ -43,7 +43,7 @@ public class UserStatisticInfoDTO extends BaseUserInfoDTO {
     /**
      * 视频播放总数
      */
-    private Long totalReadCount;
+    private Integer totalReadCount;
 
     /**
      * 是否关注当前用户
